requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
cryptography==36.0.2
Django==3.2.10
django-autocomplete-light==3.8.2
django-ckeditor==5.9.0
django-dal==1.0.0
django-el-pagination==3.3.0
django-js-asset==2.2.0
django-mail-queue==3.2.5
django-mptt==0.15.0
django-registration-redux==2.10
django-versatileimagefield==2.2
djangorestframework==3.12.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
idna==3.6
numpy==1.25.2
openpyxl==3.1.2
pandas==2.1.0
Pillow==9.0.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.5
python-magic==0.4.25
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.1.0
